voice_assistant/requirements.txt
pyyaml
google-assistant-library
google-auth-oauthlib
adapt-parser
paho-mqtt
gTTS